requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.104.1
granian==0.7.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
ruff==0.1.3
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typer==0.9.0
typing_extensions==4.8.0
uvloop==0.18.0
watchfiles==0.21.0